[ALPS02248025] PWRAP: cherry-pick pwrap upstream driver.
[Detail]
soc: mediatek: Add PMIC wrapper for MT8135 and MT8173 SoCs

This adds support for the PMIC wrapper found on MediaTek MT8135 and
MT8173 SoCs. The PMIC wrapper is found on MT6xxx SoCs aswell but these
are currently not supported.

On MediaTek MT8135, MT8173 and other SoCs the PMIC is connected via
SPI. The SPI master interface is not directly visible to the CPU, but
only through the PMIC wrapper inside the SoC. The communication between
the SoC and the PMIC can optionally be encrypted. Also a non standard
Dual IO SPI mode can be used to increase speed. The MT8135 also supports
a special feature named "IP Pairing". With IP Pairing the pins of some
SoC internal peripherals can be on the PMIC. The signals of these pins
are routed over the SPI bus using the pwrap bridge. Because of these
optional non SPI conform features the PMIC driver is not implemented as
a SPI bus master driver.

[Solution]
cherry-pick pwrap upstream driver from
commit 1f022d84bd193ec88b01c6d2cb5b6647cb20a02f ("soc: mediatek: Add PMIC wrapper for MT8135 and MT8173 SoCs").

[Feature]
N/A

MTK-Commit-Id: 07a7d71c3e8c33a47dad5ac42d254200eb4853f2

Change-Id: I09816d3472325e626c91a69e5b611c346e0aee81
CR-Id:ALPS02248025
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>