[ALPS02250657] SMI: 8163: Implement the smi driver
[Detail]
Use the SMI_VARIANT for smi common code and implement the SoC
specific configure for smi.

[Solution]
Modules will use mtk_smi_clock_on/off interface to control the
clocks and power-domain of smi common, this patch implement the
smi driver for mt8163, most of the code have been re-used from
other chips, as in the variant folder.

And some SoC specific configuration are in the files of
smi_variant_config_8163, after this patch, the clocks and power
domain manipulated functions will be workable for mt8163.

This version is the version 2 of the patch, it removes the urgly
ccflags in Makefile, get the platform information from dtsi, and
also do some code cleanups.

[Feature]
smi

[Other]
Pass the build without any warnings.

MTK-Commit-Id: 70d5a65210cd70e6a990461502a4ead9a2adccf4

Change-Id: I326e7a8724a56129eeb47774986e4997137a6452
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02250657