[ALPS02322363] clk: mt6735: Only set MTCMOS if status is changed
[Detail]
Since display MTCMOS is turned on again if system is awaken from IPOH,
smi callback is also called and the issue will happen due to callback
will close related clock in a short time. Therefore, we should check the
status before turning MTCMOS on or off.

[Solution]
Only turn MTCMOS on or off if the status is changed

[Feature] Instant Power On-Hybrid (IPO-H)

MTK-Commit-Id: abb94a9e9e59c2f16e111cd26473e3a9798e4721

Change-Id: I9ab2c5f9d29900f4693b9e2a18417e385e7dd9eb
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02322363