[ALPS02292598] SMI: 8127: Set the BWC in init_setting
[Detail]
The bandwidth control should be set in init_setting.

[Solution]
If the default value of bandwidth control have not been saved,
the register value should be saved, if it has been saved, then
the value should be restored to the register in init_setting. Or
the bandwidth control will not work in init_setting.

[Feature]
SMI

MTK-Commit-Id: 7f3e02d45bcb71a4a6bdc3f31d55e61c9cf94e4d

Change-Id: I0e289cd54031bcab4050972c72a3a4dbd158e8b4
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02292598