[ALPS02333873] mmc: core: Fix init_card in 52Mhz
[Detail] N/A

[Solution] N/A

[Feature] N/A

Suppose that we got a data crc error, and it triggers the mmc_reset.
mmc_reset will call mmc_send_status to see if HW reset was supported.
before issue CMD13, it will do retune, and if EMMC was in HS400 mode,
it will reduce frequency to 52Mhz firstly, then results in card init
was doing at 52Mhz.
The mmc_send_status was originally only done for mmc_test, should drop
it. And, rename the "eMMC hardware reset" to "Reset test", as we would
also be able to use the test for SD-cards.

MTK-Commit-Id: a20b4f14666076c8bef7df192287be5401ad2808

Change-Id: I018f63d8e270fe2888783223ebd15bc786caad59
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
Suggested-by: Adrian Hunter <adrian.hunter@intel.com>
CR-Id: ALPS02333873