[ALPS02386286] wlan: p2p: crossmountm: Devices Connection failed(2/5)
[Detail]
when device A received neg req, it will send neg resp promptly. if at
the same time, the listen timeout. it may cause the listen channel
released and cause that cant received neg confirm.

[Solution]
when supplicant sent neg resp, pass the wait time to driver, and in
driver, extended the listen interval which has ROCed before.

[Feature] CrossMount

MTK-Commit-Id: c54a6b26f5a133f90a5169583b7b71ac4b29b3f0

Change-Id: I360c7994bb58d631908ad270b29a84682c3bdf04
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02386286