[ALPS02283088] Thermal: fix a change thermal policy bug
[Detail]
When default thermal config is in use and ATM is activated,
ATM coolers are not deactivated on changing thermal policy to ht120.
The reason is that exit point of ATM coolers is not reached.
Since changing thermal config is intended, ATM coolers should be deactivated
even when their exit points not reached.

[Solution]
Set exit_threshold to 0 for deactivate the cooler then restore it back.

MTK-Commit-Id: d0d543023c7c75e3341f5394b85e6c1e0d265230

Change-Id: I20866d8a8789a1e13bc77638e38ada875e9d961e
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02283088
Feature: Thermal Management