[ALPS02433749] CCCI: add flag for CLDMA IRQ
[Detail]
We enable and disable CLDMA IRQ using kernels API.
But it has a depth counter, so if we disable twice and
enable once, the IRQ is still disabled.

[Solution]
add a atomic flag to tell whether CLDMA IRQ is disabled
or not, if already disabled, do not disable it again.

MTK-Commit-Id: 392961dd0c3d88341b90f78d512a66e67303b263

Change-Id: Ica189854260dfa8da4cba5371ff9c1cc9f07df97
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02433749
Feature: Modem Interface Driver