[ALPS02698270] mm/page_owner.c: remove unnecessary stack_trace field
Page owner uses the page_ext structure to keep meta-information for every
page in the system.  The structure also contains a field of type struct
stack_trace, page owner uses this field during invocation of the function
save_stack_trace.  It is easy to notice that keeping a copy of this
structure for every page in the system is very inefficiently in terms of
memory.

The patch removes this unnecessary field of page_ext and forces page owner
to use a stack_trace structure allocated on the stack.

MTK-Commit-Id: 7f98cd4c7622c902b880c8b641b163fd2526236a

Signed-off-by: Sergei Rogachev <rogachevsergei@gmail.com>
CR-Id: ALPS02698270
Feature: Others
Change-Id: I3fbba4e57b20fdabe0298ac2433b61447be69f7a
(cherry picked from commit d98a346f387f7b90653a02609f42e0858b22e518)