[ALPS02778999] usb20_otg_if.c: Possible buffer overflow in musb_h_in_data
[Detail]
The function musb_h_in_data accepts a buffer and a corresponding length value.
Although the code attempts to ensure that data is not written beyond the end of
the buffer, the validation currently implemented is insufficient and a buffer
overflow condition appears possible. An attacker might be able
to use this vulnerability to execute arbitrary code in the context of the
untrusted kernel, given physical access to a device.

[Solution]
Check that received+count is less than len prior to the musb_otg_read_fifo
call.

MTK-Commit-Id: 5ba9e97dc6bd262dc1853a3e17160f42e3096549

Change-Id: I2e40185e4352a7c286c9b08a4939d11ee0cd8196
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02778999
Feature: [Android Default] USB OTG