[ALPS02788976] mmc: mediatek: fix CMD21/CMD19 timeout issue
[Detail]
we did not deal with the read data of CMD21/CMD19 if there is
response CRC error of CMD21/CMD19, in this case, eMMC/SD may
still in send-data state. therefore, all of next commands cannot
get response as device is not in transfer state.

[Solution]
for resolving this issue, still need deal with the data receive
to make device back to transfer state.

MTK-Commit-Id: ea7aa9247843db5fa34900ebca6afec43d67f78d

Change-Id: I1a069bcc1db75bdcfdce704707cc9450c4147f27
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02788976
Feature: eMMC Boot Up