[ALPS02785973] Charger: Set CV in Battery Full Action
[Detail]
For JEITA load, if temperature is in normal range, the CV
setting will not be updated in JEITA state machine but in
pchr_turn_on_charging().
Since pchr_turn_on_charging() will not be called in battery
full action, we need to also update CV setting in this
condition.

[Solution]
Set CV in battery full action

MTK-Commit-Id: de26f3d690a37909307b4c899085d6ed3025f195

Change-Id: I1bc7213260d739b3c17fc43e5f41cd3aa957895d
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02785973
Feature: JEITA Battery Protection