[ALPS02807130] scp: fix unaligned abort
[Detail]
There is a shared dram buffer between scp and AP.
The original map of device memory setting is not suitable for AP access.
It results in an unexpected unaligned abort.

[Solution]
To solve this issue, replace the API with ioremap_wc().
It maps the buffer as non-cacheable and normal memory.

MTK-Commit-Id: 477fb835c71345685908e581988abe890bab4339

Change-Id: I61c57f7d08fe8d128f61636d2659ed8c57c5c933
CR-Id: ALPS02807130
Feature: Tinysys
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>