[ALPS02843372] [Google Security Patch][CVE-2016-4470]
[Detail]
 An elevation of privilege vulnerability in the kernel security
subsystem could enable a local malicious application to execute
arbitrary code within the context of the kernel. This issue is
rated as Critical due to the possibility of a local permanent
device compromise, which may require reflashing the operating
system to repair the device.

If __key_link_begin() failed then "edit" would be uninitialized.
So a fix is needed. However, on our platform the CONFIG_KEYS is
not enabled by default, so the impact is relatively lower.

Refer to:
 http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/
commit/?id=38327424b40bcebe2de92d07312c89360ac9229a

[Solution]
 1. Added a check to fix that.

MTK-Commit-Id: 50beffab293d5c0c7f577934aec05530157a2065

Change-Id: I753ec69ecf1800af0e10d103f2ddef49592ac177
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02843372
Feature: Others