[ALPS02879469] AEE: revise kernel API dump flow
[Detail] there are two processes may access the global variable of aed
driver, one is debuggerd and the other is kworker. The working flow that
we expected is as following:
1. kworker generates one oops one time and wait for child debuggerd to
handle it.
2. child debuggerd get the oops and output one DB as a result, then
notify kworker by complete() function.
3. kworker clean up the related message of this oops.
4. kworker check if there is other oops need to be generated, and repeat
step 1.2.3.4 if yes.

However, debuggerd may have chance to do step 2 again while kworker was
still doing setp 3 or the setp 1 of next oops, which lead to serious
problem.

Example 1, there is chance that debuggerd fork child process to handle
one oops during the time between oops is passed to .lastlog and the .msg
successfully being assigned. Refer to CR ALPS02820034.

Example 2, while debuggerd is in the complete() function the kworker may
calling init_completion, which free the lock that complete() still
holding. Refer to CR alps02879469.

[Solution] add a variable to control debuggerd to fork child debuggerd
when it is safe.

MTK-Commit-Id: f3894d694b90ef7411cfc69ad9fce85b836f5c23

Change-Id: I5d1a1b6865dcb593dcbb9b72ed53493c6d19186a
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02879469
Feature: Android Exception Engine(AEE)