[ALPS02924520] mem: switch to lowpower when init
Switch the DRAM setting to lowpower mode in DCS init function.

Previous implementation assumes the DRAM can be in lowpower
mode even the kernel DCS driver is disabled. So we need addition
APIs to handle this situation and it leads to a bad design.

This patch assumes that the DRAM is in normal mode before
its initialization. The DCS H/W should be disabled when the
DCS kernel driver is off. The assumption works with platforms
which do not have the DCS H/W.

MTK-Commit-Id: bbb8b1e7173532f16f9ef5f326adeb87fa386874

Change-Id: I5bd98ee5fcea882c3bcde2eee8b5ddfe92c955c9
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02924520
Feature: DCS (Dynamic Channel Switch)