[ALPS02923224] UPSTREAM: KEYS: Fix ASN.1 indefinite length object parsing
[Detail]
This fixes CVE-2016-0758.

In the ASN.1 decoder, when the length field of an ASN.1 value is extracted,
it isnt validated against the remaining amount of data before being added
to the cursor.  With a sufficiently large size indicated, the check:

	datalen - dp < 2

may then fail due to integer overflow.

Fix this by checking the length indicated against the amount of remaining
data in both places a definite length is determined.

Whilst were at it, make the following changes:

 (1) Check the maximum size of extended length does not exceed the capacity
     of the variable its being stored in (len) rather than the type that
     variable is assumed to be (size_t).

 (2) Compare the EOC tag to the symbolic constant ASN1_EOC rather than the
     integer 0.

 (3) To reduce confusion, move the initialisation of len outside of:

	for (len = 0; n > 0; n--) {

     since it doesnt have anything to do with the loop counter n.

Signed-off-by: David Howells <dhowells@redhat.com>
Reviewed-by: Mimi Zohar <zohar@linux.vnet.ibm.com>
Acked-by: David Woodhouse <David.Woodhouse@intel.com>
Acked-by: Peter Jones <pjones@redhat.com>

MTK-Commit-Id: 5e50994c8ae8fc200cda74675862b0ff292710cb

Change-Id: I3b90ff7109d89a414f397e8a91c06569da2cb446
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02923224
Feature: [Android Default] SELinux, SEAndroid, and SE-MTK