[ALPS02923264] Binder: Prevent heap address of binder service from potential leak
[Detail]
The interaction between the kernel /dev/binder and the usermode
Parcel.cpp means that when a Binder object is passed as
BINDER_TYPE_BINDER or BINDER_TYPE_WEAK_BINDER, a pointer to that object
(in the server process) is leaked to the client process as the cookie
value. This leads to a leak of a heap address in many of the privileged
Binder services, including system_server.

The fix is designed to zero out the Binder pointer and cookie before
sending it to the client process.

Bug: A-30768347

MTK-Commit-Id: c48af2d58e7444a4960922743ec1b5d8c46c9988

Change-Id: I4d2380b10c42a4273c96c92f53171ab115d54dad
CR-Id: ALPS02923264
Feature: Binder
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>