[ALPS02875181] wlan: Bug fix of repeating oid complete
[Detail]
The wlanOidSetCountryCode is the oid handler of kalIoctl,
if there is any oid cmd to send in wlanOidSetCountryCode,
the return value of wlanOidSetCountryCode must be WLAN_STATUS_PENDING.
In rlmDomainSendCmd, the second parameter is TRUE,
means the cmd we send is oid. We found two oid cmd,
so we complete oid repeatly.

[Solution]
Modify the second parameter of rlmDomainSendCmd to FALSE.

MTK-Commit-Id: 015b0feef63aa991fbed15a46dfa51d5c2d57caa

Change-Id: Iaf1f8fd0adcc018ca83aa6499e142f4a5096b092
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS02875181
Feature: [Android Default] Wi-Fi Driver for SoC