[ALPS03411241] binder: remove binder_debug_no_lock mechanism
(from https://patchwork.kernel.org/patch/9817811/)

With the global lock, there was a mechanism to acceess
binder driver debugging information with the global
lock disabled to debug deadlocks or other issues.
This mechanism is rarely (if ever) used anymore
and wasnt needed during the development of
fine-grained locking in the binder driver.
Removing it.

MTK-Commit-Id: 0833f88053c2b9c7135369eaaf410cd15237a642

Change-Id: Ie1cf45748cefa433607a97c2ef322e7906efe0f7
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Money Kang <Money.Kang@mediatek.com>
CR-Id: ALPS03411241
Feature: binder