[ALPS03411241] binder: avoid race conditions when enqueuing txn
(from https://patchwork.kernel.org/patch/9817813/)

Currently, the transaction complete work item is queued
after the transaction. This means that it is possible
for the transaction to be handled and a reply to be
enqueued in the current thread before the transaction
complete is enqueued, which violates the protocol
with userspace who may not expect the transaction
complete. Fixed by always enqueing the transaction
complete first.

Also, once the transaction is enqueued, it is unsafe
to access since it might be freed. Currently,
t->flags is accessed to determine whether a sync
wake is needed. Changed to access tr->flags
instead.

MTK-Commit-Id: d5cbac9c77a14b27b36f9956a284cea48aa26b15

Change-Id: I6c01566e167a39cf17c9027c3817618182e56975
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Money Kang <Money.Kang@mediatek.com>
CR-Id: ALPS03411241
Feature: binder