[ALPS03411241] binder: use atomic for transaction_log index
(from https://patchwork.kernel.org/patch/9817807/)

The log->next index for the transaction log was
not protected when incremented. This led to a
case where log->next++ resulted in an index
larger than ARRAY_SIZE(log->entry) and eventually
a bad access to memory.

Fixed by making the log index an atomic64 and
converting to an array by using "% ARRAY_SIZE(log->entry)"

Also added "complete" field to the log entry which is
written last to tell the print code whether the
entry is complete

MTK-Commit-Id: 219f28f3e91f07b59f7d04a8433a0e9603f15f7c

Bug: 62038227
Test: tested manually
Change-Id: I1bb1c1a332a6ac458a626f5bedd05022b56b91f2
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Money Kang <Money.Kang@mediatek.com>
CR-Id: ALPS03411241
Feature: binder