[ALPS03411241] binder:protect agst stale ptr in print_binder_transaction
(from https://patchwork.kernel.org/patch/9817761/)

When printing transactions there were several race conditions
that could cause a stale pointer to be deferenced. Fixed by
reading the pointer once and using it if valid (which is
safe). The transaction buffer also needed protection via proc
lock, so it is only printed if we are holding the correct lock.

MTK-Commit-Id: 087daf6e62d3b1d5060ad6b53866896e200c8412

Bug: 36650912
Test: tested manually
Change-Id: I78240f99cc1a070d70a841c0d84d4306e2fd528d
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Money Kang <Money.Kang@mediatek.com>
CR-Id: ALPS03411241
Feature: binder