[ALPS03292700] ASoC: deep buffer music seek has pop noise
[Detail]
music seek has pop noise in deep buffer playback.

[Solution]
The original solution is to extend the buffer size of
deep buffer kernel size. The cons of this workaround is
increasing the cost of HW SRAM. Another option is to
clear the kernel buffer in ack callback function,
after ALSA copy.

MTK-Commit-Id: 4c1f6d9cc74f15b69a6828cdebf15f53285a4e44

Change-Id: Ie3d737ecfb1c957376d6fc3c90713f7e12e38b7f
Signed-off-by: Money Kang <Money.Kang@mediatek.com>
CR-Id: ALPS03292700
Feature: Phone Sound