[ALPS03475340] ANDROID: mnt: Fix next_descendent
next_descendent did not properly handle the case
where the initial mount had no slaves. In this case,
we would look for the next slave, but since dont
have a master, the check for wrapping around to the
start of the list will always fail. Instead, we check
for this case, and ensure that we end the iteration
when we come back to the root.

MTK-Commit-Id: 45c62ea7599947066d538bb2a22ab42df91f49fe

Signed-off-by: Daniel Rosenberg <drosen@google.com>
Bug: 62094374
Change-Id: I43dfcee041aa3730cb4b9a1161418974ef84812e
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Feature: others
CR-Id: ALPS03475340
(cherry picked from commit 63fb3f929445b8766ef50423811e2ff1822e1e66)