[ALPS03526613] wlan: sync rlm params with FW too many
[Detail]
When driver receives beacon or probe response, it will
try to sync the received rlm params with FW no matter
the IEs in the beacon are same as the pervious.

[Solution]
Before we send the syncing command to FW, check if the
params are different. If so, send it; Otherwise, ignore
the sending procedure.

MTK-Commit-Id: 44c50b9b57564f85b1ba02de00931bac890d88e0

Change-Id: I0d2384cb632c464ab24a181bb7db744845be6f50
Signed-off-by: Wayne-TW Chen <wayne-tw.chen@mediatek.com>
CR-Id: ALPS03526613
Feature: Wifi Driver