[ALPS03373988] touch: Fix a bug of writing out of bounds
[Detail]
According to the definition in mstart_drv_mutual_test.h,
ANA4_MUTUAL_CSUB_NUMBER should be equal to 408,
which causes the arrary szDbBusTxData Writed out of bounds,
and the macro definition in the header file also does not match the corresponding
comment.
[Solution]
Modify the macro definition in the header file.

MTK-Commit-Id: 46bcf9ab2359ccb3c68c851979accd4fef4a58eb

Change-Id: I4eb4d69da28fce394afdebdfce3111faf0b99de7
Signed-off-by: ZH Chen <zh.chen@mediatek.com>
CR-Id: ALPS03373988
Feature: [Android Default] CapTouch