[ALPS03575657] disp: use static DC buf in GMO project
[Detail]
We used to use dynamic decouple buffer in GMO project to
save memory. However, its possible that we cannot create
decouple buffer when we need it because of low memory. This
situation would cause there is no buffer for WDMA and trigger
M4U translation fault

[Solution]
Create a static decouple buffer for decouple mode.

MTK-Commit-Id: 6a9bfa2d6b9e0f8c73184c5785bcd316cba0329d

Change-Id: Ia93ca0d29621e3bbb94358534ff447ea52298d34
Signed-off-by: mtk13530 <elon.hsu@mediatek.com>
CR-Id: ALPS03575657
Feature: MTBF(*)