[ALPS03505581] vfs: add support for a lazytime mount option
Add a new mount option which enables a new "lazytime" mode.  This mode
causes atime, mtime, and ctime updates to only be made to the
in-memory version of the inode.  The on-disk times will only get
updated when (a) if the inode needs to be updated for some non-time
related change, (b) if userspace calls fsync(), syncfs() or sync(), or
(c) just before an undeleted inode is evicted from memory.

This is OK according to POSIX because there are no guarantees after a
crash unless userspace explicitly requests via a fsync(2) call.

For workloads which feature a large number of random write to a
preallocated file, the lazytime mount option significantly reduces
writes to the inode table.  The repeated 4k writes to a single block
will result in undesirable stress on flash devices and SMR disk
drives.  Even on conventional HDDs, the repeated writes to the inode
table block will trigger Adjacent Track Interference (ATI) remediation
latencies, which very negatively impact long tail latencies --- which
is a very big deal for web serving tiers (for example).

MTK-Commit-Id: 72cd0db73076f6f562608ca400e07cd972c4ec6f

Change-Id: Ic686429b0e325e157f09a52fe0da2efd32570ae8
CR-Id: ALPS03505581
Feature: vfs
Signed-off-by: Chin-Han Chen <chin-han.chen@mediatek.com>
(cherry picked from commit fd0eda546a85c362ad426656f1863c6f6dc31414)