[ALPS03673038] msg2238a: Protect against kernel memory corruption
We now verify the user-provided nCount prior to our call to
copy_from_user(), to assure the user cannot write out of bounds
of a global array.

We switch from the hardcoded "1024" to "sizeof(_gDebugBug)" in
the memset() also to future-proof this code from possible buffer
size changes.

MTK-Commit-Id: 250243a4c68bdc1322b5294fe66fd9c8afecaf08

Test: Build and flashed a dual sim gobo device, and confirmed touchscreen is fine.
Bug: 70511487
CR-Id: ALPS03673038
Feature: AGO Google Patch
Change-Id: I147ab1e480db27e654161fac732c251dff11c5c5
Signed-off-by: Chin-Han Chen <chin-han.chen@mediatek.com>