[ALPS03673038] msg2238a: Protect against kernel memory corruption
When the user directly provides a size for writing to a kernel
stack buffer, we validate it wont overflow, and return with
an error if it would.

When the user provides a data structure with the size of data
to read out, we check the size and (silently) truncate it if
it would have read beyond our buffer.

We use a shared define for a couple buffer sizes, to avoid
hard-coding 1024 throughout the code and risking values getting
out of sync in future code changes.

MTK-Commit-Id: bdb668309ff2f74b10ee55bdca7f730abe60f0d7

Test: Build and flashed a dual sim gobo device, and confirmed touchscreen is fine.
Bug: 70511798
CR-Id: ALPS03673038
Feature: AGO Google Patch
Change-Id: I090cc6bb8fda542e63909b2d1651dbeff8f54912
Signed-off-by: Chin-Han Chen <chin-han.chen@mediatek.com>