[ALPS03727915] Audio: google security patch
[Detail]
In the MIDI driver global variables in certain ioctl calls
are not correctly protected with locks.
This could lead to double-free or use-after-free errors in the kernel.
[Solution]
The fix is designed to fix the locking in the affected functions.

MTK-Commit-Id: 2d95fa1b5bcf02b0b5b84422494802aa12df38f4

Change-Id: I5afd7e46076e2c75e8f3b9a43e95fa28487dcf78
Signed-off-by: chipeng chang <chipeng.chang@mediatek.com>
CR-Id: ALPS03727915
Feature: Others
(cherry picked from commit b199b2a3f4202ab735e846554392f825a6996b36)