[ALPS03805834] Wlan: ID Vulnerability in NL80211 driver
[Detail]
[Google Security Patch][CVE-2017-15837]ID Vulnerability in NL80211
driver

Define a policy for packet pattern attributes in order to fix a
potential read over the end of the buffer during nla_get_u32()
of the NL80211_PKTPAT_OFFSET attribute.

Note that the data there can always be read due to SKB allocation
(with alignment and struct skb_shared_info at the end), but the
data might be uninitialized. This could be used to leak some data
from uninitialized vmalloc() memory, but most drivers dont allow
an offset (so youd just get -EINVAL if the data is non-zero) or
just allow it with a fixed value - 100 or 128 bytes, so anything
above that would get -EINVAL. With brcmfmac the limit is 1500 so
(at least) one byte could be obtained.

[Solution]
Apply patch (from kernel-4.4)

MTK-Commit-Id: 060c3889ed204fdc1379ff1b4f4fe62a7720b91c

Change-Id: I005f9c46329b8f6e9f6e1f7e60c263c10a1708ec
Signed-off-by: Simfex Chiu <simfex.chiu@mediatek.com>
CR-Id: ALPS03805834
Feature: Others