[ALPS03913680] ASN.1: fix out-of-bounds read
[Detail]
Including 2 patches

1. [PATCH] ASN.1: fix out-of-bounds read when parsing indefinite length
item
In asn1_ber_decoder(), indefinitely-sized ASN.1 items were being passed
to the action functions before their lengths had been computed, using
the bogus length of 0x80 (ASN1_INDEFINITE_LENGTH).  This resulted in
reading data past the end of the input buffer, when given a specially
crafted message.

Fix it by rearranging the code so that the indefinite length is resolved
before the action is called.

This bug was originally found by fuzzing the X.509 parser in userspace
using libFuzzer from the LLVM project.

2. [PATCH] KEYS: fix out-of-bounds read during ASN.1 parsing
syzkaller with KASAN reported an out-of-bounds read in
asn1_ber_decoder().  It can be reproduced by the following command,
assuming CONFIG_X509_CERTIFICATE_PARSER=y and CONFIG_KASAN=y:

    keyctl add asymmetric desc $x30x30 @s

The bug is that the length of an ASN.1 data value isnt validated in the
case where it is encoded using the short form, causing the decoder to
read past the end of the input buffer.  Fix it by validating the length.

MTK-Commit-Id: 91de62760f08622e49692f62d8b956c43b56c24d

Change-Id: Ic306af959a5474e5e167552dd9655a187c15b50d
Signed-off-by: Jades Shih <jades.shih@mediatek.com>
CR-Id: ALPS03913680
Feature: [Module]Official Kernel Patch